refactor(custom-icons): use 'usvg' crate
- Use `usvg` crate to parse and transform the paths instead of the `svg`
  and `svg_path_ops` crates.
- Simplify the `svg_parser` using the `usvg` crate and its internal
  `tiny_skia_path` for path operations.

+++ src/lib.rs
@@ -372,7 +372,7 @@ pub(crate) struct ChosenGlyph {
 #[derive(Clone, Debug, Serialize)]
 pub(crate) struct SvgIcon {
     path: String,
-    width: f64,
+    width: f32,
 }
 
 fn parse_fonts() -> BTreeMap<String, Font> {
